poem-gemini/requirements.txt
python-dotenv==1.0.1
google-genai==1.7.0
